Handle wrapped sequence numbers safely
Agent-Logs-Url: https://github.com/zoghby-ctrl/Multiplayer-Pong-CPP/sessions/e4dc1fd2-3946-422b-9dec-28404248391e

+++ Server/src/main.cpp
@@ -75,7 +75,7 @@ int main() {
 
         Packet statePacket{};
         statePacket.header.type = PacketType::State;
-        statePacket.header.seq = global_state.tick;
+        statePacket.header.seq = static_cast<uint32_t>(global_state.tick);
         statePacket.payload.state = global_state;
 
         SendPacketToAll(statePacket);
